Add simple loki dep
